+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.2.2] - 2018-09-14
+### Fixed
+- Fixed bug that occurred when no para.yml configuration file exists.
+- Fixed bug that prevented to add a new project when no group has been configured in the configuration file.
+
 ## [2.2.1] - 2018-03-17
 ### Changed
 - Improved the list of available plugins. Now the latest stable release will be preferred.